+++ includes/admin/class-wta-settings.php
@@ -18,6 +18,12 @@ class WTA_Settings {
 		register_setting( 'wta_data_import_settings_group', 'wta_github_countries_url', array( 'sanitize_callback' => 'esc_url_raw' ) );
 		register_setting( 'wta_data_import_settings_group', 'wta_github_states_url', array( 'sanitize_callback' => 'esc_url_raw' ) );
 		register_setting( 'wta_data_import_settings_group', 'wta_github_cities_url', array( 'sanitize_callback' => 'esc_url_raw' ) );
+		
+		// Performance Settings
+		register_setting( 'wta_data_import_settings_group', 'wta_concurrent_batches', array( 
+			'sanitize_callback' => 'absint',
+			'default' => 10
+		) );
 
 		// AI Settings
 		register_setting( 'wta_ai_settings', 'wta_openai_api_key', array( 'sanitize_callback' => 'sanitize_text_field' ) );